#782 pipelines not starting

See merge request mlreef/mlreef!892

+++ backend/mlreef-rest/src/main/kotlin/com/mlreef/rest/api/v1/ExperimentsController.kt
@@ -10,8 +10,8 @@ import com.mlreef.rest.FileLocation
 import com.mlreef.rest.Person
 import com.mlreef.rest.api.v1.dto.DataProcessorInstanceDto
 import com.mlreef.rest.api.v1.dto.ExperimentDto
-import com.mlreef.rest.api.v1.dto.PipelineJobInfoDto
 import com.mlreef.rest.api.v1.dto.FileLocationDto
+import com.mlreef.rest.api.v1.dto.PipelineJobInfoDto
 import com.mlreef.rest.api.v1.dto.toDto
 import com.mlreef.rest.exceptions.ConflictException
 import com.mlreef.rest.exceptions.ErrorCode
@@ -41,7 +41,7 @@ import javax.validation.constraints.NotEmpty
 
 @RestController
 @RequestMapping("/api/v1/data-projects/{dataProjectId}/experiments")
-class ExperimentsController(
+internal class ExperimentsController(
     val service: ExperimentService,
     val dataProjectService: ProjectService<DataProject>,
     val pipelineService: PipelineService,
@@ -50,65 +50,68 @@ class ExperimentsController(
 ) {
     private val log: Logger = Logger.getLogger(ExperimentsController::class.simpleName)
 
-    private fun beforeGetExperiment(experimentId: UUID): Experiment {
-        return experimentRepository.findByIdOrNull(experimentId)
-            ?: throw NotFoundException(ErrorCode.NotFound, "Experiment with id $experimentId not found")
-    }
-
     private fun beforeGetExperiment(dataProjectId: UUID, idOrNumber: String): Experiment {
-        val number: Int? = tryOrNull { idOrNumber.toInt() }
-        val id: UUID? = tryOrNull { UUID.fromString(idOrNumber) }
+        val number: Int? = idOrNumber.toIntOrNull()
+        val id: UUID? =
+            try {
+                { UUID.fromString(idOrNumber) }.invoke()
+            } catch (e: Exception) {
+                null
+            }
 
         val found = when {
             number != null -> experimentRepository.findOneByDataProjectIdAndNumber(dataProjectId, number)
             id != null -> experimentRepository.findOneByDataProjectIdAndId(dataProjectId, id)
-            else -> throw NotFoundException(ErrorCode.NotFound, "Experiment not found: '$idOrNumber' not a valid id/number")
+            else -> throw NotFoundException(ErrorCode.NotFound, "Experiment not found: '$idOrNumber' neither a valid UUID nor number")
         }
         return found ?: throw NotFoundException(ErrorCode.NotFound, "Experiment with id/number $idOrNumber not found")
     }
 
     @GetMapping
     @PreAuthorize("canViewProject(#dataProjectId)")
-    fun getAllExperiments(@PathVariable dataProjectId: UUID): List<ExperimentDto> {
-        val experiments: List<Experiment> = experimentRepository.findAllByDataProjectId(dataProjectId).toList()
-        return experiments.map(Experiment::toDto)
-    }
+    fun getAllExperiments(@PathVariable dataProjectId: UUID): List<ExperimentDto> =
+        experimentRepository.findAllByDataProjectId(dataProjectId)
+            .map { it.toDto() }
+
 
     @GetMapping("/{idOrNumber}")
     @PreAuthorize("canViewProject(#dataProjectId)")
-    fun getExperiment(@PathVariable dataProjectId: UUID, @PathVariable idOrNumber: String): ExperimentDto {
-        val experiment = beforeGetExperiment(dataProjectId, idOrNumber)
-        return experiment.toDto()
-    }
+    fun getExperiment(@PathVariable dataProjectId: UUID, @PathVariable idOrNumber: String): ExperimentDto =
+        beforeGetExperiment(dataProjectId, idOrNumber).toDto()
 
-    fun <R> tryOrNull(func: () -> R): R? = try {
-        func.invoke()
-    } catch (e: Exception) {
-        null
-    }
 
     @GetMapping("/{idOrNumber}/info")
     @PreAuthorize("canViewProject(#dataProjectId)")
-    fun getExperimentMetrics(@PathVariable dataProjectId: UUID, @PathVariable idOrNumber: String): PipelineJobInfoDto {
-        val experiment = beforeGetExperiment(dataProjectId, idOrNumber)
-        return experiment.pipelineJobInfo?.toDto()
+    fun getExperimentMetrics(@PathVariable dataProjectId: UUID, @PathVariable idOrNumber: String): PipelineJobInfoDto =
+        beforeGetExperiment(dataProjectId, idOrNumber)
+            .pipelineJobInfo
+            ?.toDto()
             ?: throw NotFoundException(ErrorCode.NotFound, "Experiment does not have a PipelineJobInfo (yet)")
-    }
+
 
     @GetMapping("/{idOrNumber}/mlreef-file", produces = [MediaType.TEXT_PLAIN_VALUE])
     @PreAuthorize("canViewProject(#dataProjectId)")
-    fun getExperimentYaml(@PathVariable dataProjectId: UUID, @PathVariable idOrNumber: String, account: Account): String {
+    fun getExperimentYaml(
+        @PathVariable dataProjectId: UUID,
+        @PathVariable idOrNumber: String,
+        account: Account
+    ): String {
         val experiment = beforeGetExperiment(dataProjectId, idOrNumber)
-        return service.createExperimentFile(experiment = experiment, author = account, secret = experiment.pipelineJobInfo?.secret
-            ?: "***censored***")
+        return service.createExperimentFile(
+            experiment = experiment,
+            author = account,
+            secret = experiment.pipelineJobInfo?.secret ?: "***censored***"
+        )
     }
 
     @PostMapping("/{idOrNumber}/start")
     @PreAuthorize("hasAccessToProject(#dataProjectId, 'DEVELOPER')")
-    fun startExperiment(@PathVariable dataProjectId: UUID,
-                        @PathVariable idOrNumber: String,
-                        account: Account,
-                        userToken: TokenDetails): PipelineJobInfoDto {
+    fun startExperiment(
+        @PathVariable dataProjectId: UUID,
+        @PathVariable idOrNumber: String,
+        account: Account,
+        userToken: TokenDetails,
+    ): PipelineJobInfoDto {
         val dataProject = dataProjectService.getProjectById(dataProjectId)
             ?: throw ProjectNotFoundException(projectId = dataProjectId)
 
@@ -122,9 +125,9 @@ class ExperimentsController(
         val pipelineJobInfo = pipelineService.createStartGitlabPipeline(userToken = userToken.accessToken, projectGitlabId = dataProject.gitlabId,
             targetBranch = experiment.targetBranch, fileContent = fileContent, sourceBranch = experiment.sourceBranch, secret = secret)
 
-        val experimentWithPipeline = service.savePipelineInfo(experiment, pipelineJobInfo)
-
-        return experimentWithPipeline.pipelineJobInfo!!.toDto()
+        return service.savePipelineInfo(experiment, pipelineJobInfo)
+            .pipelineJobInfo!!
+            .toDto()
     }
 
     @PostMapping
@@ -140,8 +143,7 @@ class ExperimentsController(
         log.info(experimentCreateRequest.toString())
 
         val slug = Slugs.toSlug(experimentCreateRequest.slug)
-        val findAllByDataProjectIdAndSlug = experimentRepository.findOneByDataProjectIdAndSlug(dataProjectId, slug)
-        findAllByDataProjectIdAndSlug?.let {
+        if (experimentRepository.findOneByDataProjectIdAndSlug(dataProjectId, slug) != null) {
             throw ConflictException(ErrorCode.ExperimentSlugAlreadyInUse, "Duplicate! There exists already an Experiment in this DataProject for this slug: $slug !")
         }
 
@@ -162,9 +164,6 @@ class ExperimentsController(
         }
 
 
-        val inputFiles = experimentCreateRequest.inputFiles.map {
-            FileLocation.fromDto(it.location, it.locationType)
-        }
         val newExperiment = service.createExperiment(
             authorId = person.id,
             dataProjectId = dataProject.id,
@@ -174,19 +173,16 @@ class ExperimentsController(
             sourceBranch = experimentCreateRequest.sourceBranch,
             targetBranch = experimentCreateRequest.targetBranch,
             postProcessors = postProcessors,
-            inputFiles = inputFiles,
-            processorInstance = processorInstance
+            inputFiles = experimentCreateRequest.inputFiles
+                .map { FileLocation.fromDto(it.location, it.locationType) },
+            processorInstance = processorInstance,
         )
 
         val persisted = experimentRepository.findOneByDataProjectIdAndId(dataProjectId, newExperiment.id)!!
         return persisted.toDto()
     }
 
-    // Please avoid fanboy-usage of kotlin expression functions! They are nice, and advised for short expressions, which are *single* expression or can be concateneted simpliy!
-    // If you need "let", it is actually a sign, that is not a "natural" single expression, but a forced one. Please keep it readable and respects the maintainers
 
-    // also "let" inverses the flow of reading the code. Code is beautiful when it is easy and readable,
-    // not just because the amount of forced usage of expression functions is maximized, nobody in the style guides suggests that
     @Deprecated("This is a workaround solution and should be removed ASAP ") //TODO //FIXME
     @PostMapping("/{id}/cancel")
     @PreAuthorize("hasAccessToProject(#dataProjectId, 'DEVELOPER')")
@@ -194,7 +190,12 @@ class ExperimentsController(
     fun updateExperimentStatus(
         @PathVariable dataProjectId: UUID,
         @PathVariable id: UUID
-    ) = experimentRepository.save(beforeGetExperiment(id).copy(status = ExperimentStatus.CANCELED)).toDto()
+    ) = experimentRepository.findByIdOrNull(id)
+        ?.copy(status = ExperimentStatus.CANCELED)
+        ?.let { experimentRepository.save(it) }
+        ?.toDto()
+        ?: throw NotFoundException(ErrorCode.NotFound, "Experiment with id $id not found")
+
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
@@ -203,13 +204,14 @@ class ExperimentsController(
         @PathVariable dataProjectId: UUID,
         @PathVariable id: UUID
     ) {
-        //Kotlin style guide: "Prefer using an expression body for functions with the body consisting of a *single* expression", and please not for Unit, that is misleading
-        beforeGetExperiment(id)
-        experimentRepository.deleteById(id)
+        if (experimentRepository.findByIdOrNull(id) != null) {
+            experimentRepository.deleteById(id)
+        }
     }
 }
 
-class ExperimentCreateRequest(
+
+internal class ExperimentCreateRequest(
     val dataInstanceId: UUID?,
     @NotEmpty val slug: String = "",
     @NotEmpty val name: String,
